refactor: Remove index in components/Buttons, Change import statement
