--- src/components/Buttons/index.js
@@ -1,5 +1,0 @@
-import DefaultButton from "./DefaultButton";
-import SmallButton from "./SmallButton";
-import SquareButton from "./SquareButton";
-
-export { DefaultButton, SmallButton, SquareButton };
