+++ src/components/Modals/FailModalContainer.js
@@ -3,7 +3,8 @@ import React, { useState } from "react";
 import { useNavigation } from "@react-navigation/native";
 import PropTypes from "prop-types";
 
-import { ButtonModal, SmallButton } from "..";
+import { ButtonModal } from "..";
+import SmallButton from "../Buttons/SmallButton";
 import { navigations, colors } from "../../common/constants";
 
 const FailModalContainer = ({ onRetryCurrentStage }) => {
