+++ src/components/Modals/InputModalContainer.js
@@ -5,7 +5,8 @@ import { Keyboard } from "react-native";
 import uuid from "react-native-uuid";
 import PropTypes from "prop-types";
 
-import { SmallButton, InputModal } from "..";
+import { InputModal } from "..";
+import SmallButton from "../Buttons/SmallButton";
 import { navigations, colors } from "../../common/constants";
 import GameContext from "../../context/GameContext";
 import {
